requirements.txt
dotenv
firecrawl-py
fpdf
tqdm
asyncio
json
datetime
numpy
matplotlib
fpdf
anthropic
openai